:bug: Fixed inaccurate Button text color & unmanaged rgb colors
The contrast problem with Button text color has been fixed. Now the colors are accurately generated to some extend. When inputing rgb colors, they are not correctly handle causing the builder to not generate contrasting color accurately. Now the rgb inputs are converted into hexadecimal colors and can be used in the whole code.